+++ talk-main/src/main/java/com/locker/config/jwt/JwtAuthenticationFilter.java
@@ -27,13 +27,6 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
     private final JwtBlacklistService blacklistService;
 
     @Override
-    protected boolean shouldNotFilter(HttpServletRequest request) {
-        String path = request.getServletPath();
-        return path.startsWith("/oauth2/")
-                || path.startsWith("/login/oauth2/");
-    }
-
-    @Override
     protected void doFilterInternal(HttpServletRequest request,
                                     HttpServletResponse response,
                                     FilterChain chain
