+++ talk-main/src/main/java/com/locker/config/security/OAuth2DelegateConfig.java
@@ -1,0 +1,14 @@
+package com.locker.config.security;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
+
+@Configuration
+public class OAuth2DelegateConfig {
+
+    @Bean
+    public DefaultOAuth2UserService defaultOAuth2UserService() {
+        return new DefaultOAuth2UserService();
+    }
+}
